# docs/openapi/FlightService.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FlightService",
    "version": "1.0"
  },
  "paths": {
    "/api/Flights/airlines/{airlineId}": {
      "get": {
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "airlineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Flights/airports": {
      "get": {
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "departureCity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destinationCity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Flights/flightListings/{flightNumber}": {
      "get": {
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "flightNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Flights/flightlistings": {
      "get": {
        "tags": [
          "Flights"
        ],
        "parameters": [
          {
            "name": "departureCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destinationCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "travelDate",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": { }
}